remove redundant pub crate
